No more warnings in VGA driver

+++ include/metodo/hal/vga.h
@@ -9,7 +9,7 @@ static void VgaDisplayScroll(void);
 void VgaDisplayChar(char c);
 void VgaDisplaySpot(uint8_t s, uint8_t row, uint8_t col);
 void VgaDisplayHideCursor();
-void VgaDisplayCursorPosition(int _row, int _col);
+void VgaDisplayCursorPosition(uint8_t _row, uint8_t _col);
 void VgaInitDisplay(void);
 void VgaDisableDisplay(void);
 int VgaIsDisplayOn(void);
